+++ src/modules/meal/meal.controller.ts
@@ -19,7 +19,7 @@ import { AuthenticationGuard } from '../../guards/authentication.guard';
 import { CreationGuard } from '../../guards/creation.guard';
 import { EditionGuard } from '../../guards/edition.guard';
 import { DeletionGuard } from '../../guards/deletion.guard';
-import { DetailedMealWithTranslatedIngredients, GetMealsQueryType } from './meal.types';
+import { DetailedMealWithTranslations, GetMealsQueryType } from './meal.types';
 import { IngredientName, MealType } from '../../common/enums';
 import { MealQueryValidationPipe } from '../../pipes/meal-query-validation.pipe';
 import { TranslationService } from '../translation/translation.service';
@@ -48,11 +48,12 @@ export class MealController {
 
     @Get('/:id/details')
     @HttpCode(200)
-    public async getMealDetails(@Param('id') id: string, @Headers('accept-language') lang: Language): Promise<DetailedMealWithTranslatedIngredients> {
+    public async getMealDetails(@Param('id') id: string, @Headers('accept-language') lang: Language): Promise<DetailedMealWithTranslations> {
         const meal = await this.mealService.getMealDetails(id);
         const ingredients = await this.translationService.translateIngredients(meal.ingredients, lang);
+        const recipe = await this.translationService.translateRecipe(meal.recipeSections, lang);
 
-        return { meal, ingredients };
+        return { meal, ingredients, recipe };
     }
 
     @Post('/create')