+++ src/modules/translation/translation.service.spec.ts
@@ -3,6 +3,7 @@ import { TranslationService } from './translation.service';
 import { translate } from 'google-translate-api-x';
 import { MealIngredient } from '../ingredient/ingredient.types';
 import { TranslatedIngredient } from './translation.types';
+import { MealRecipeSections } from '../meal/meal.types';
 
 jest.mock('google-translate-api-x', () => ({
     translate: jest.fn((text) => {
@@ -27,6 +28,50 @@ describe('TranslationService', () => {
 
     it('should be defined', () => {
         expect(translationService).toBeDefined();
+    });
+
+    describe('translateRecipe', () => {
+        afterEach(() => {
+            jest.clearAllMocks();
+        });
+
+        it('should translate array of recipe sections', async () => {
+            const sections: MealRecipeSections = [
+                {
+                    name: 'X',
+                    steps: [
+                        {
+                            number: 1,
+                            step: 'Y'
+                        },
+                        {
+                            number: 2,
+                            step: 'Z'
+                        }
+                    ]
+                }
+            ];
+            const expectedResult: MealRecipeSections = [
+                {
+                    name: 'X',
+                    steps: [
+                        {
+                            number: 1,
+                            step: 'mocked translation of Y'
+                        },
+                        {
+                            number: 2,
+                            step: 'mocked translation of Z'
+                        }
+                    ]
+                }
+            ];
+
+            const translatedRecipeSections = await translationService.translateRecipe(sections, 'pl');
+
+            expect(translate).toHaveBeenCalledTimes(sections[0].steps.length);
+            expect(translatedRecipeSections).toStrictEqual(expectedResult);
+        });
     });
 
     describe('translateIngredients', () => {