+++ src/modules/translation/translation.service.ts
@@ -3,9 +3,35 @@ import { translate } from 'google-translate-api-x';
 import { MealIngredient } from '../ingredient/ingredient.types';
 import { Language } from '../../common/types';
 import { TranslatedIngredient } from './translation.types';
+import { MealRecipeSection, MealRecipeSections, MealRecipeStep } from '../meal/meal.types';
 
 @Injectable()
 export class TranslationService {
+
+    async translateRecipe(recipeSections: MealRecipeSections, targetLanguage?: Language): Promise<MealRecipeSections> {
+        const language = targetLanguage ?? 'en';
+
+        return await Promise.all(
+            recipeSections.map(
+                section => this._translateRecipeSection(section, language)
+            )
+        );
+    }
+
+    async _translateRecipeSection(recipeSection: MealRecipeSection, targetLanguage: Language): Promise<MealRecipeSection> {
+        const { name, steps } = recipeSection;
+
+        const translatedSteps: string[] = await Promise.all(steps.map(step => this.translate(step.step, targetLanguage)));
+        const completedSteps: MealRecipeStep[] = translatedSteps.map((step, idx) => ({
+            number: idx + 1,
+            step: step
+        }));
+
+        return {
+            name,
+            steps: completedSteps
+        };
+    }
 
     async translateIngredients(ingredients: MealIngredient[], targetLanguage?: Language): Promise<TranslatedIngredient[]> {
         const translatedIngredients = ingredients.map(async (ingredient) => {