+++ CHANGELOG.md
@@ -6,11 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.37.0] - 2024-08-20
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): `google-translate-api-x` library.
-- [Marek Kurańda](https://github.com/mjkuranda): `TranslateModule` to translate all ingredient expressions.
+- [Marek Kurańda](https://github.com/mjkuranda): `TranslateModule` to translate all ingredient expressions and recipe sections.
 - [Marek Kurańda](https://github.com/mjkuranda): `Accept-Language` header in `GET /meals` to define language for an incoming client.
+- [Marek Kurańda](https://github.com/mjkuranda): Unit tests for `MealController`.
 
 ### Changed
-- [Marek Kurańda](https://github.com/mjkuranda): `GET /meal/:id/details` return type to `DetailedMealWithTranslatedIngredients`.
+- [Marek Kurańda](https://github.com/mjkuranda): `GET /meal/:id/details` return type to `DetailedMealWithTranslations`.
 
 ### Removed
 - [Marek Kurańda](https://github.com/mjkuranda): `IngredientUnit` enum.